refactor: move page sections
