+++ src/components/layout/page-sections/lister-section/rr-lister-section.ts
@@ -1,29 +1,37 @@
 /**
- * RegelRecht Simple Section Component (Lit + TypeScript)
+ * RegelRecht Lister Section Component (Lit + TypeScript)
  *
- * A basic section with responsive container padding.
+ * A section with a CSS grid layout for displaying card collections.
  *
- * @element rr-simple-section
+ * @element rr-lister-section
  * @attr {string} container - Container size: 'sm' | 'md' | 'lg' (default: 'md')
  *
- * @slot - Default slot for content
+ * @slot - Default slot for grid items
  *
  * @csspart section - The section container
  * @csspart body - The section body
- * @csspart main - The main content area
+ * @csspart main - The main content area (grid)
  *
  * @cssprop --rr-section-gap - Override gap between sections
  */
 
-import { LitElement, html } from 'lit';
+import { LitElement, html, css } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import { sectionSharedStyles } from './section-shared-styles.js';
+import { sectionSharedStyles } from '../section-shared-styles.js';
 
 type Container = 'sm' | 'md' | 'lg';
 
-@customElement('rr-simple-section')
-export class RRSimpleSection extends LitElement {
-  static override styles = [sectionSharedStyles];
+@customElement('rr-lister-section')
+export class RRListerSection extends LitElement {
+  static override styles = [
+    sectionSharedStyles,
+    css`
+      .section__main {
+        display: grid;
+        grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+      }
+    `,
+  ];
 
   @property({ type: String, reflect: true })
   container: Container = 'md';
@@ -43,6 +51,6 @@ export class RRSimpleSection extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    'rr-simple-section': RRSimpleSection;
+    'rr-lister-section': RRListerSection;
   }
 }
