+++ src/components/layout/page-sections/two-thirds-one-third-section/rr-two-thirds-one-third-section.stories.js
@@ -1,9 +1,9 @@
 import { html } from 'lit';
-import './rr-one-half-one-half-section.ts';
+import './rr-two-thirds-one-third-section.ts';
 
 export default {
-  title: 'Components/Layout/One Half One Half Section',
-  component: 'rr-one-half-one-half-section',
+  title: 'Components/Layout/Page Sections/Two Thirds One Third Section',
+  component: 'rr-two-thirds-one-third-section',
   tags: ['autodocs'],
   parameters: {
   },
@@ -18,15 +18,15 @@ export default {
 
 const sampleContent = html`
   <div style="padding: 24px; background: #f1f5f9; border-radius: 8px;">
-    <h2 style="margin: 0 0 8px 0;">Left Content (1/2)</h2>
-    <p style="margin: 0;">This is the left half content area.</p>
+    <h2 style="margin: 0 0 8px 0;">Main Content (2/3)</h2>
+    <p style="margin: 0;">This is the main content area taking two thirds of the width.</p>
   </div>
 `;
 
 const sampleAside = html`
   <div style="padding: 24px; background: #e2e8f0; border-radius: 8px;">
-    <h3 style="margin: 0 0 8px 0;">Right Content (1/2)</h3>
-    <p style="margin: 0;">This is the right half content area.</p>
+    <h3 style="margin: 0 0 8px 0;">Aside (1/3)</h3>
+    <p style="margin: 0;">This is the aside content.</p>
   </div>
 `;
 
@@ -35,9 +35,9 @@ export const Default = {
     container: 'md',
   },
   render: (args) => html`
-    <rr-one-half-one-half-section container=${args.container}>
+    <rr-two-thirds-one-third-section container=${args.container}>
       ${sampleContent}
       <div slot="aside">${sampleAside}</div>
-    </rr-one-half-one-half-section>
+    </rr-two-thirds-one-third-section>
   `,
 };
