+++ src/components/layout/page-sections/two-thirds-one-third-section/rr-two-thirds-one-third-section.ts
@@ -1,14 +1,14 @@
 /**
- * RegelRecht One Third Two Thirds Section Component (Lit + TypeScript)
+ * RegelRecht Two Thirds One Third Section Component (Lit + TypeScript)
  *
- * A split layout section with a 1/3 aside on the left and 2/3 main content on the right.
+ * A split layout section with 2/3 main content on the left and a 1/3 aside on the right.
  * Stacks vertically on small screens.
  *
- * @element rr-one-third-two-thirds-section
+ * @element rr-two-thirds-one-third-section
  * @attr {string} container - Container size: 'sm' | 'md' | 'lg' (default: 'md')
  *
  * @slot - Default slot for main content
- * @slot aside - Slot for aside content (left column)
+ * @slot aside - Slot for aside content (right column)
  *
  * @csspart section - The section container
  * @csspart body - The section body
@@ -20,12 +20,12 @@
 
 import { LitElement, html, css } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import { sectionSharedStyles } from './section-shared-styles.js';
+import { sectionSharedStyles } from '../section-shared-styles.js';
 
 type Container = 'sm' | 'md' | 'lg';
 
-@customElement('rr-one-third-two-thirds-section')
-export class RROneThirdTwoThirdsSection extends LitElement {
+@customElement('rr-two-thirds-one-third-section')
+export class RRTwoThirdsOneThirdSection extends LitElement {
   static override styles = [
     sectionSharedStyles,
     css`
@@ -58,12 +58,12 @@ export class RROneThirdTwoThirdsSection extends LitElement {
     return html`
       <section class="section" part="section">
         <div class="section__body" part="body">
-          <aside class="section__aside" part="aside">
-            <slot name="aside"></slot>
-          </aside>
           <div class="section__main" part="main">
             <slot></slot>
           </div>
+          <aside class="section__aside" part="aside">
+            <slot name="aside"></slot>
+          </aside>
         </div>
       </section>
     `;
@@ -72,6 +72,6 @@ export class RROneThirdTwoThirdsSection extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    'rr-one-third-two-thirds-section': RROneThirdTwoThirdsSection;
+    'rr-two-thirds-one-third-section': RRTwoThirdsOneThirdSection;
   }
 }
